+++ EltGeo.h
@@ -40,7 +40,7 @@ public:
     //
 
     const string getNom();
-    virtual void deplacer(int dx, int dy) =0;
+    virtual void deplacer(long dx, long dy) =0;
     virtual string description() = 0 ;
     // Mode d'emploi : deplace l'element geometrique selon la translation de vecteur(x0,y0)
     //
